requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
blas=1.0=mkl
ca-certificates=2020.7.22=0
certifi=2020.6.20=py37_0
cloudpickle=1.6.0=py_0
cycler=0.10.0=py37_0
cytoolz=0.11.0=py37haf1e3a3_0
dask-core=2.28.0=py_0
decorator=4.4.2=py_0
freetype=2.10.2=ha233b18_0
imageio=2.9.0=py_0
intel-openmp=2019.4=233
jpeg=9b=he5867d9_2
kiwisolver=1.2.0=py37h04f5b5a_0
lcms2=2.11=h92f6f08_0
libcxx=10.0.0=1
libedit=3.1.20191231=h1de35cc_1
libffi=3.3=hb1e8313_2
libgfortran=3.0.1=h93005f0_2
libpng=1.6.37=ha441bb4_0
libtiff=4.1.0=hcb84e12_1
lz4-c=1.9.2=hb1e8313_1
matplotlib=3.3.1=0
matplotlib-base=3.3.1=py37h181983e_0
mkl=2019.4=233
mkl-service=2.3.0=py37hfbe908c_0
mkl_fft=1.2.0=py37hc64f4ea_0
mkl_random=1.1.1=py37h959d312_0
ncurses=6.2=h0a44026_1
networkx=2.5=py_0
ninja=1.10.1=py37h879752b_0
numpy=1.19.1=py37h3b9f5b6_0
numpy-base=1.19.1=py37hcfb5961_0
olefile=0.46=py37_0
openssl=1.1.1h=haf1e3a3_0
pillow=7.2.0=py37ha54b6ba_0
pip=20.2.3=py37_0
pyparsing=2.4.7=py_0
python=3.7.9=h26836e1_0
python-dateutil=2.8.1=py_0
pytorch=1.6.0=py3.7_0
pywavelets=1.1.1=py37h1de35cc_0
pyyaml=5.3.1=py37haf1e3a3_1
readline=8.0=h1de35cc_0
scikit-image=0.16.2=py37h6c726b0_0
scipy=1.5.2=py37h912ce22_0
setuptools=49.6.0=py37_1
six=1.15.0=py_0
sqlite=3.33.0=hffcf06c_0
tk=8.6.10=hb0a8c7a_0
toolz=0.11.1=py_0
torchvision=0.7.0=py37_cpu
tornado=6.0.4=py37h1de35cc_1
wheel=0.35.1=py_0
xz=5.2.5=h1de35cc_0
yaml=0.2.5=haf1e3a3_0
zlib=1.2.11=h1de35cc_3
zstd=1.4.5=h41d2c2f_0
